Created log in controller to create users
